argument_dummy

argument_dummy

argument_dummy(object, phase, pt_alpha, pt_clr, pt_clrp, pt_clrsp, pt_size)

Arguments

object
A valid cypro object.

phase
Character value. The ordinal value referring to the phase of interest (e.g. 'first', 'second' etc.). referring to the phase of interest. If set to NULL takes the phase denoted as default with adjustDefault().

Ignored if the experiment design contains only one phase.

pt_alpha
Numeric value. Specifies the degree of transparency of all points.

pt_clr
Character value. Specifies the color of all points.

pt_clrp
The color palette to be used if the specified variable displayed by color is categorical/discrete. Run validColorPalettes() to see valid input.

pt_clrsp
The color spectrum to be used if the specified variable displayed by color is continuous. Run validColorSpectra() to see valid input.

pt_size
Numeric value. Specifies the size of all points.

across
Character value, character vector or NULL (see details for more). Specifies the grouping variable of interest.

Use getGroupingVariableNames() to obtain all variable names that group the cells of your experiment in a certain manner.

across_subset
Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

add_on_list
A list of ggplot2-add-ons that are supposed to be integrated in the visualization process.

cell_ids
Character vector. Denotes the cell ids of interest.

clrp
Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust
Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

clrsp
Character value. Specifies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

discrete_feature
Character value. Specifies the name of the grouping variable of interest. Use getGroupingOptions() to obtain all valid input options.

display_cols
Logical value. If set to TRUE columns are used additionally to display the results.

display_facets
Logical value. If set to TRUE the plot is split via ggplot2::facet_wrap() such that each variable gets it's own subplot.

display_legend
Logical value. If set to TRUE a legend is displayed.

display_line
Logical value. If set to TRUE a line is used additionally to display the results.

display_points
Logical value. If set to TRUE points are used additionally to display the results.

display_smooth
Logical value. If set to TRUE a smoothed line is displayed to emphasize the overall trend of the data. Use the smooth_* arguments to additionally adjust they way the trend is plotted.

display_title
Logical value. If set to TRUE an informative title is displayed.

force
Logical value. Needs to be set to TRUE if you want to overwrite an already existing set up or already existing results.

image
Numeric value. The well-image of interest.

linesize
Numeric value. Denotes the size of the lines drawn.

linetype
Character value. Valid options are 'solid', 'twodash', 'longdash', 'dotted' and 'dotdash'.

method_aggl
Character vector (or value see details for more.) Denotes the agglomeration method(s) of interest according to which the existing distance matrices are agglomerated to hierarchical trees.

Use validAgglomerationMethods() to obtain all valid input options.

method_corr
Character value. Denotes the correlation method of interest. Either 'pearson' or 'spearman'.

method_dist
Character vector (or value see details for more.) Denotes the distance method(s) of interest (e.g. 'euclidean' or 'manhattan').

Use validDistanceMethods() to obtain all valid input options.

method_kmeans
Character vector (or value see details for more.) Denotes the algorithms of interest. Defaults to 'Hartigan-Wong'.

Use validKmeansMethods() to obtain all valid input options.

method_outlier
Character vector. Specifies the method/algorithm of interest.

Use validOutlierDetectionMethods() to obtain all valid input options.

method_pam
Character vector (or value see details for more.) Denotes the algorithm of interest. Valid input options are 'euclidean' and 'manhattan'.

n_cells
Numeric calue. Determines the number of cells that are randomly chosen from every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.

overwrite
Logical value. Must be set to TRUE in case of overlapping variable names.

(Note that overwriting stat variables leads to all analysis progress of affected variable sets being discarded.)

pretty_names
Logical. If set to TRUE the function attempts to convert the concisely named variables into more aesthetically pleasing ones.

relevel
Logical value. If set to TRUE the input order of across_subset determines the order in which the groups of interest are displayed. Groups that are not included are dropped which affects the colors with which they are displayed.

scales, space, ncol, nrow
Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

simplify
Logical. If set to TRUE the output list is simplified to a vector if possible. If set to FALSE a list is returned.

smooth
Logical. If set to TRUE the values are smoothed.

smooth_clr
Character value. Denotes the color of the smoothed line.

smooth_method
Character value. Denotes the model type used to display the line. Defaults to 'lm' (linear model). Given to argument method of function ggplot2::geom_smooth().

smooth_se
Logical. If set to TRUE the standard error will be displayed.

smooth_size
Numeric value. Denotes the size of the smoothed line.

smooth_span
NUmeric value. Denotes the smoothing span used.

well
Character value. The well of interest (e.g. 'A1', 'B12')

well_plate
Character value. The name of the well plate of interest. Valid inputs can be obtained via getWellPlateNames().

with_cluster
Logical. If set to TRUE the discrete variables of the meta data slot are added to the output data.frame.

with_meta
Logical. If set to TRUE the discrete variables of the meta data slot are added to the output data.frame.

with_stats
Logical. If set to TRUE the numeric variables of the stat data slots are added tot he output data.frame.

variable_set
Character value. Denotes the variable set of interest. Use getVariableSetNames() to obtain all names of currently stored variable sets in your object.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)