Obtain cypros clustering objects

Obtain cypros clustering objects

getHclustConv(object, variable_set, phase = NULL, with_data = TRUE)

getHclustObject(object, variable_set, phase = NULL, with_data = TRUE)

getKmeansConv(object, variable_set, phase = NULL, with_data = TRUE)

getKmeansObject(object, variable_set, phase = NULL, with_data = TRUE)

getPamConv(object, variable_set, phase = NULL, with_data = TRUE)

getPamObject(object, variable_set, phase = NULL, with_data = TRUE)

Arguments

object
A valid cypro object.

variable_set
Character value. Denotes the variable set of interest. Use getVariableSetNames() to obtain all names of currently stored variable sets in your object.

phase
Character value. The ordinal value referring to the phase of interest (e.g. 'first', 'second' etc.). referring to the phase of interest. If set to NULL takes the phase denoted as default with adjustDefault().

Ignored if the experiment design contains only one phase.

Value

An S4 object of 'hclust_conv', 'kmeans_conv' or 'pam_conv'.