Object initiation: Step 2

Opens an interactive application in which the data to be analyzed is loaded.

loadDataFile() assumes that the data exist in one .csv, .xls, .xlsx or .txt file and that this data table contains variables/columns that identify the well plate, the well and the region of interest.

loadDataFiles() assumes that the data tables exist for every well-region of interest combination (every stack of images). Here you can assign a directory to every well plate leading to a folder in which the files are placed. Successful reading of the files requires them to be named according to the syntax: well_roi.filetype, e.g A1_1.xlsx, B12_4.csv

loadDataFile(object)

loadDataFiles(object)

Arguments

object
A valid cypro object.

Value

An updated cypro object.