Plot scaled cell migration

Visualizes the cells migration in a typical migration plot. Scales the cell's x- and y-coordinates such that all cell's migration paths start from the same position.

plotAllTracks(
  object,
  across = "condition",
  across_subset = NULL,
  relevel = NULL,
  time_subset = NULL,
  phase = NULL,
  n_cells = 100,
  color_by = across,
  linetype = "solid",
  linesize = 0.75,
  clrp = "milo",
  clrp_adjust = NULL,
  display_annotation = TRUE,
  set_seed = NULL,
  ...,
  verbose = TRUE
)

Arguments

object
A valid cypro object.

across
Character value, character vector or NULL (see details for more). Specifies the grouping variable of interest.

Use getGroupingVariableNames() to obtain all variable names that group the cells of your experiment in a certain manner.

across_subset
Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

relevel
Logical value. If set to TRUE the input order of across_subset determines the order in which the groups of interest are displayed. Groups that are not included are dropped which affects the choice of color.

phase
Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. 'first', 'second' etc.). referring to the phase of interest or 'all'. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with adjustDefault().

Ignored if the experiment design contains only one phase.

n_cells
Numeric calue. Determines the number of cells that are randomly chosen from every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.

linetype
Character value. Valid options are 'solid', 'twodash', 'longdash', 'dotted' and 'dotdash'.

linesize
Numeric value. Denotes the size of the lines drawn.

clrp
Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust
Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

display_annotation
Logical. If set to TRUE the number of cells (input for argument n_cells) is displayed in the right upper corner of each plot.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

A ggplot.