Visualize changes of cell characteristics over time

Uses the design of a heatmap to visualize feature dynamics over time across a grouping variable of interest.

plotTimeHeatmap(
  object,
  variable,
  across = "condition",
  across_subset = NULL,
  phase = "all",
  n_cells = 100,
  color = NA,
  colors = viridis::viridis(15),
  smooth = TRUE,
  smooth_span = 0.25,
  arrange_rows = "maxima",
  verbose = TRUE,
  ...,
  in_shiny = FALSE
)

Arguments

object
A valid cypro object.

across
Character value or NULL. Specifies the grouping variable of interest.

Use getGroupingOptions() to obtain all variable names that group the cells of your experiment in a certain manner.

across_subset
Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

phase
Character value. The ordinal value referring to the phase of interest (e.g. 'first', 'second' etc.). referring to the phase of interest. If set to NULL takes the phase denoted as default with adjustDefault().

Ignored if the experiment design contains only one phase.

n_cells
Numeric calue. Determines the number of cells that are randomly chosen from every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.

colors
A vector of colors that is used as the heatmap's color spectrum.

smooth
Logical. If set to TRUE the values are smoothed.

smooth_span
NUmeric value. Denotes the smoothing span used.

arrange_rows
Character value. Either 'maxima' or 'none'.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...
Additional arguments given to pheatmap::pheatmap().