Plot cell activity over time

Visualizes the percentage of active cells over time.

plotVelocityLineplot(
  object,
  across = "condition",
  across_subset = NULL,
  phase = NULL,
  threshold = NULL,
  linesize = 1,
  smooth = TRUE,
  smooth_span = 0.25,
  smooth_se = FALSE,
  clrp = "milo",
  verbose = TRUE,
  ...,
  in_shiny = FALSE
)

Arguments

object
A valid cypro object.

across
Character value or NULL. Specifies the grouping variable of interest.

Use getGroupingOptions() to obtain all variable names that group the cells of your experiment in a certain manner.

across_subset
Character vector or NULL. Specifies the particular groups of interest the grouping variable specified in argument across contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in with a '-'. (Saves writing if there are more groups you are interested in than groups you are not interested in.)

Use getGroupNames() to obtain all valid input options.

phase
Character value. The ordinal value referring to the phase of interest (e.g. 'first', 'second' etc.). referring to the phase of interest. If set to NULL takes the phase denoted as default with adjustDefault().

Ignored if the experiment design contains only one phase.

threshold
Numeric value or NULL. If set to NULL (the default) the threshold to consider a cell 'active' is equal to base::mean(speed) + base::sd(speed)

linesize
Numeric value. Denotes the size of the lines drawn.

smooth
Logical. If set to TRUE the values are smoothed.

smooth_span
NUmeric value. Denotes the smoothing span used.

smooth_se
Logical. If set to TRUE the standard error will be displayed.

clrp
Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)