Plot the well plate set up

Plot the well plate set up

plotWellPlate(
  object,
  phase = NULL,
  well_plates = NULL,
  color_by = "condition",
  summarize_with = "median",
  plot_type = "well",
  well_size = 13.5,
  display_border = TRUE,
  border_clr = "black",
  border_size = 1,
  display_edge = TRUE,
  edge_clr = "black",
  display_labels = TRUE,
  alpha = 0.9,
  clrp = "milo",
  clrp_adjust = NULL,
  clrsp = "Blues",
  ncol = NULL,
  nrow = NULL,
  ...
)

Arguments

object
A valid cypro object.

phase
Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. 'first', 'second' etc.). referring to the phase of interest or 'all'. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with adjustDefault().

Ignored if the experiment design contains only one phase.

plot_type
Character value. Either 'well' or 'tile'. Affects the geometry with which the wells are displayed. (Either with ggplot2::geom_point() or with ggplot2::geom_tile()).

well_size
Numeric value. Denotes the size with which each well is plotted if argument plot_type is set to 'well'.

display_border
Logical value. If set to TRUE a line is drawn around the geometric objects displaying the wells.

border_clr
Character value. Denotes the color of the wells borders.

border_size
Numeric value. Denotes the thickness of the wells borders.

display_edge
Logical value. If set to TRUE ggforce::geom_mark_rect() is used to display the edges of the well plate.

edge_clr
Character value. Denotes the color of the well plates edge.

clrp
Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust
Named character vector or NULL. If character, it adjusts the color palette that is used to represent the groups. Names of the input vector must refer to the group and the respective named element denotes the color with which to represent the group.

clrsp
Character value. Specifies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

ncol
Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

nrow
Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

Value

A ggplot.